More work on self-hosting + projects.

+++ reset.sh
@@ -2,20 +2,23 @@
 
 set -x
 
-CLUSTER_NAME="development"
 SSH_ENDPOINT_HOSTNAME="atlantis"
 SSH_ENDPOINT_DOMAIN_NAME="ancapistan.io"
 TEST_DOMAIN="ancapistan.casa"
+CLUSTER_NAME="development"
 
 export LXD_VM_NAME="${TEST_DOMAIN//./-}"
 
-lxc delete --force www-"$LXD_VM_NAME"
-lxc delete --force btcpay-"$LXD_VM_NAME"
-lxc delete --force sovereign-stack
-lxc delete --force sovereign-stack-base
+if [ -n "$TEST_DOMAIN" ]; then
+    lxc delete --force www-"$LXD_VM_NAME"
+    lxc delete --force btcpay-"$LXD_VM_NAME"
+    lxc delete --force sovereign-stack
+    lxc delete --force sovereign-stack-base
 
-lxc profile delete www-"$LXD_VM_NAME"
-lxc profile delete btcpay-"$LXD_VM_NAME"
+    lxc profile delete www-"$LXD_VM_NAME"
+    lxc profile delete btcpay-"$LXD_VM_NAME"
+fi
+
 lxc profile delete sovereign-stack
 
 lxc image rm sovereign-stack-base
@@ -30,6 +33,7 @@ lxc remote remove "$CLUSTER_NAME"
 
 source "$HOME/.bashrc"
 
-./cluster.sh create "$CLUSTER_NAME" "$SSH_ENDPOINT_HOSTNAME.$SSH_ENDPOINT_DOMAIN_NAME" --data-plane-interface=enp89s0
+./cluster.sh create "$CLUSTER_NAME" "$SSH_ENDPOINT_HOSTNAME.$SSH_ENDPOINT_DOMAIN_NAME" 
+#--data-plane-interface=enp89s0
 
 #./deploy.sh
